Remove the last flaky heredoc.t test 'heredoc-tmpfile-8'.
It fails randomly, and I suspect there is something different and funny
going on in test-harness with particular perl versions (but I may as
well be totally wrong about this).
My suspicion arose after testing directly this particular test in the
shell itself, and never failing that way. So this actually works In The
Real World(tm), but sometimes the perly test fails.
It fails so that it won't output the second or the third 'hi', after
successful outputs of the first and the forth 'hi'. The one that is
missing from the output (2# or 3#), seems random (I know, it just seems
random...)