Sync. r1.40 var.c
"Fri Dec 12 05:00:55 2014 UTC
Range check the value of the base for typeset -i base x=y.  Instead of
erroring for values outside of the range switch to base 10 to match the
behaviour of ksh93.  As we have a smaller digit alphabet than ksh93 the
accepted range is smaller (2-36) than ksh93 (2-64).  The other form
of setting a base, typeset -i x=base#y already has a range check that
errors (as ksh93 also does for that syntax).
Fixes a crash found with the afl fuzzer."

/jsg@/