main.c: sync r1.55
"date: 2015/02/09 09:09:30;  author: jsg;  state: Exp;  lines: +8 -1;
commitid: uzKTVAHRqy49RYm8;
If we hit multiple errors while unwinding we'll end up running code that
assumes a pointer is valid when it has been free'd.

This is convoluted as ksh has it's own allocator and uses long jumps.
Set the pointer to NULL after the quitenv() call in unwind() in case we
later hit a long jump in unwind().

Found with afl a while back.

ok tedu@ deraadt@"

/jsg@/