ksh can run with pledge "stdio rpath wpath cpath getpw fattr proc exec tty"
ksh can run with pledge "stdio rpath wpath cpath getpw fattr proc exec tty"
if the mknod builtin is disabled.  It looks like a lot of abilities, but
hey, this is a shell.  can't open sockets or do other nasty stuff though.
(we'll leave the mknod builtin enabled on the install media for now; there
is work happening to regain the MAKEDEV performance in a different way)
discussions with otto & millert in particular
Members:
	Makefile:1.29->1.30
	c_sh.c:1.49->1.50
	main.c:1.60->1.61

Date: 2015/10/10 02:10:07
Author: deraadt
Patchset: 555

I want to derive as little as possible from upstream, so merge the
changes, but define MKNOD in the Makefile and don't actually compile the
pledge(2) call.