Add allocarray(), an overflow-safe allocation function.
Add allocarray(), an overflow-safe allocation function.
We avoided reallocation support because it demands more fancy footwork
to deal with the prepended link struct.
This has been on my mind for a while, and a 2010 security review of mksh
by the Android security team's Chris Palmer suggested it.
ok nicm@. Also discussed with millert@ and tedu@.
Members:
	alloc.c:1.8->1.9

Date: 2015/10/16 05:17:56
Author: mmcc
Patchset: 565