When ERREXIT (set -e) is in effect, don't disable it after a
When ERREXIT (set -e) is in effect, don't disable it after a
short-circuited `&&` or `||` chain in later iterations of `for`
loops.  It should be disabled after a short-circuited `&&` or `||`
chain only in the *final* iteration of loops (as well as in
conditionals, which was already working).  From Kartik Agaram.
Members:
	exec.c:1.66->1.67

Date: 2016/12/11 18:44:48
Author: millert
Patchset: 680