The fact is simple: Subshells have the same PPID as the current shell.
The fact is simple: Subshells have the same PPID as the current shell.
What is not so simple is squinting hard enough to conclude that POSIX
actually says that. I'm still not sure how to do that.
Problem with the shell manual pointed out by "Choose a display name".
ok otto, halex, jmc
Members:
	sh.1:1.142->1.143

Date: 2017/06/01 22:37:38
Author: tb
Patchset: 694