Remove expensive pointer check in afree()
Remove expensive pointer check in afree()
The check added in rev 1.8 walks the whole freelist to catch cases where
an unknown pointer is passed to afree(); but it can't catch cases
whether the struct link has been corrupted by an invalid memory write.
And it becomes very expensive when you have lots of items in an area
(for example with a huge HISTSIZE).
Discussed with & ok millert@ tb@
Members:
	alloc.c:INITIAL->1.17

Date: 2017/08/15 19:57:57
Author: jca
Patchset: 709