Include hostname in shell prompts by default
Include hostname in shell prompts by default
With tmux, ssh and vmd, we tend to open shells on many different hosts
simultaneously and the default prompts '$ ' and '# ' for {,k}sh as well
as  '% ' and '# ' for csh become dangerous: it's very easy to issue a
command on the wrong host.
This can easily be avoided by displaying the hostname in the prompt.
Everything beyond "hostname{$,#,%} " is going to be a matter of taste,
so we left it at that. If you use an FQDN, only the first part (the
output of 'hostname -s') will be printed.
requested by and ok deraadt; mostly positive feedback many
ok anton, brynet, bcallah and others
Members:
	lex.c:1.72->1.73
	main.c:1.84->1.85

Date: 2017/12/12 01:18:58
Author: tb
Patchset: 736