The hack for /dev/fd support on systems without it was removed a
The hack for /dev/fd support on systems without it was removed a
long time ago.  Update the comments and replace test_stat() calls
with stat() since test_stat() now just calls stat() directly.  Also
rename the "mode" parameter to "amode" in test_eaccess() to match
access(2) and make it clear that this is the access mode and not
the file mode.  OK jca@
Members:
	c_test.c:1.23->1.24

Date: 2017/12/26 20:10:31
Author: millert
Patchset: 741