Include "id" in pledge (for setres[ug]id, setgroups) if the shell is
Include "id" in pledge (for setres[ug]id, setgroups) if the shell is
privileged and remove it when dropping privileges (set +p), setting a
flag to make sure we don't do it again.
ok deraadt millert
Members:
	main.c:1.93->1.94
	misc.c:1.70->1.71

Date: 2018/11/16 07:41:58
Author: nicm
Patchset: 780

I actually removed the 'issetugid' and 'dropped_privileges' stuff from
misc.c because it wouldn't make sense here -- we're not even using the
pledge stuff in main.c, it's just there to ease occasional merging.