Add support for set -o pipefail
Date: 2020-07-07 10:33:58 +0000
Author: jca <jca@openbsd.org>
Original commit id: cbb0b321fa62f17a1e02f67ceb4d6f5f7bdc606b

Add support for set -o pipefail

With the pipefail option set, the exit status of a pipeline is 0 if all
commands succeed, or the return status of the rightmost command that
fails.  This can help stronger error checking, but is not a silver
bullet.  For example, commands will exhibit a non-zero exit status if
they're killed by a SIGPIPE when writing to a pipe.  Yet pipefail was
considered useful enough to be included in the next POSIX standard.

This implementation remembers the value of the pipefail option when
a pipeline is started, as described as option 1) in

  https://www.austingroupbugs.net/view.php?id=789#c4102

Requested by ajacoutot@, ok millert@