In VI command line editing mode, make the behavi..
Date: 2025-04-25 18:28:33 +0000
Author: schwarze <schwarze@openbsd.org>
Original commit id: 82b01731a2e4ef11c51876e8c29a6060d73d8014

In VI command line editing mode, make the behaviour of "1P" consistent with the behaviour of "2P" such that both back up to the first byte of the last character inserted.  As a side benefit, that also makes "[n]P" work in the presence of UTF-8 characters.

According to pascal@, the new behaviour of "P" is conistent with ksh93
and all historical versions of ksh he tested on sdf.org, so the deviant
behaviour of "1P" was likely not intentional, but a bug in pdksh.
According to Walter Alejandro Iglesias <wai at roquesor dot com>,
the new behaviour is also conistent with vim and bash.

OK pascal@